[YUNIKORN-2954] Remove so-called modular sidebar
Remove the modular remote sidebar functionality, as it goes against the
open philosophy of Apache YuniKorn development.

This reverts commit 21ebe5d9b1bbfc2e1eea8bbb8440d2a8bd38af4e.
