--- src/app/components/allocations-drawer/allocations-drawer.component.scss
@@ -1,154 +1,0 @@
-/**
- * Licensed to the Apache Software Foundation (ASF) under one
- * or more contributor license agreements.  See the NOTICE file
- * distributed with this work for additional information
- * regarding copyright ownership.  The ASF licenses this file
- * to you under the Apache License, Version 2.0 (the
- * "License"); you may not use this file except in compliance
- * with the License.  You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
- 
-.mat-drawer-container {
-  min-width: 430px;
-  width: 55%;
-  height: calc(100vh - 60px);
-  background: transparent;
-  pointer-events: none;
-  position: absolute;
-  right: 0;
-  bottom: 0;
-  .mat-drawer {
-    pointer-events: auto;
-    width: 100%;
-    .content {
-      padding: 0 10px;
-    }
-  }
-
-  .full-row-height {
-    border: 1px solid red;
-  }
-
-  .mat-mdc-header-cell {
-    font-size: 15px;
-    font-weight: bold;
-    color: #666;
-  }
-  .mat-mdc-cell {
-    color: #333;
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-    .mat-res-ul{
-      padding: 0;
-      margin: 0;
-      .mat-res-li{
-        list-style-type: none;
-      }
-    }
-    .mat-toggle-more {
-      display: block;
-      color: #a1a3b7;
-      padding: 5px 45px 0 0;
-      &:hover {
-        color: #8d00d4;
-      }
-    }
-  }
-  .mat-mdc-row {
-    &:hover {
-      background: #cccccc;
-      cursor: pointer;
-    }
-    &.selected-row {
-      background: #303d54;
-      .mat-mdc-cell {
-        color: #fff;
-        .mat-toggle-more {
-          color: #b8bbff;
-          &:hover {
-            color: #b871dc;
-          }
-        }
-      }
-    }
-  }
-  .mat-mdc-header-cell.indicator-icon,
-  .mat-mdc-cell.indicator-icon {
-    max-width: 40px;
-    font-size: 18px;
-    margin-left: 10px;
-  }
-
-  .close-btn {
-    float: right;
-    font-size: 1.2em;
-    cursor: pointer;
-    padding-right: 5px;
-    &:hover {
-      color: #f44336;
-    }
-  }
-  .copy-btn {
-    font-size: 1em;
-    cursor: pointer;
-    padding-left: 5px;
-  }
-  .header {
-    margin: 20px;
-    font-weight: 100;
-    font-size: 1em;
-  }
-  .content {
-    border-top: 1px solid #e1e1e1;
-  }
-  .item-wrapper {
-    .left-item {
-      text-align: right;
-    }
-    .left-item,
-    .right-item {
-      width: 50%;
-      padding: 6px;
-      padding-right: 0;
-      color: #666;
-    }
-    .right-item {
-      font-weight: 600;
-    }
-  }
-  .app-link {
-    text-decoration: none;
-    color: #666;
-    &:hover {
-      text-decoration: underline;
-    }
-  }
-
-  .row {
-    font-size: 0.9em;
-  }
-
-  .even-row {
-    background: #eee;
-  }
-
-  .ellipsis {
-    white-space: nowrap;
-    overflow: hidden;
-    text-overflow: ellipsis;
-    display: inline-block;
-  }
-
-  .row {
-    min-height: unset;
-  }
-}
