+++ src/app/components/apps-view/apps-view.component.spec.ts
@@ -31,15 +31,8 @@ import { By, HAMMER_LOADER } from '@angular/platform-browser';
 import { NoopAnimationsModule } from '@angular/platform-browser/animations';
 import { RouterTestingModule } from '@angular/router/testing';
 import { AppInfo } from '@app/models/app-info.model';
-import { EnvconfigService } from '@app/services/envconfig/envconfig.service';
 import { SchedulerService } from '@app/services/scheduler/scheduler.service';
-import { MatChipsModule } from '@angular/material/chips';
-
-import {
-  MockEnvconfigService,
-  MockNgxSpinnerService,
-  MockSchedulerService,
-} from '@app/testing/mocks';
+import { MockNgxSpinnerService, MockSchedulerService } from '@app/testing/mocks';
 import { NgxSpinnerService } from 'ngx-spinner';
 import { of } from 'rxjs';
 
@@ -64,20 +57,15 @@ describe('AppsViewComponent', () => {
         MatTooltipModule,
         MatSelectModule,
         MatSidenavModule,
-        MatChipsModule,
       ],
       providers: [
         { provide: SchedulerService, useValue: MockSchedulerService },
         { provide: NgxSpinnerService, useValue: MockNgxSpinnerService },
         { provide: HAMMER_LOADER, useValue: () => new Promise(() => {}) },
-        { provide: EnvconfigService, useValue: MockEnvconfigService },
       ],
     }).compileComponents();
     fixture = TestBed.createComponent(AppsViewComponent);
     component = fixture.componentInstance;
-    spyOn(component, 'initializeSidebarComponent').and.callFake(
-      (b = null) => new Promise(() => {})
-    );
     fixture.detectChanges();
   });
 
@@ -85,7 +73,7 @@ describe('AppsViewComponent', () => {
     expect(component).toBeTruthy();
   });
 
-  it('should have usedResource and pendingResource column with detailToggle OFF', () => {
+  it('should have usedResource and pendingResource column', () => {
     let service: SchedulerService;
     service = TestBed.inject(SchedulerService);
     let appInfo = new AppInfo(
@@ -102,45 +90,22 @@ describe('AppsViewComponent', () => {
     );
     spyOn(service, 'fetchAppList').and.returnValue(of([appInfo]));
     component.fetchAppListForPartitionAndQueue('default', 'root');
+    component.toggle();
     fixture.detectChanges();
     const debugEl: DebugElement = fixture.debugElement;
     expect(
-      debugEl.query(By.css('[data-test="Memory: 500.0 KB,CPU: 10,pods: 1"]')).nativeElement
-        .innerText
-    ).toContain('Memory: 500.0 KB\nCPU: 10');
-    expect(
-      debugEl.query(By.css('[data-test="Memory: 0.0 bytes,CPU: 0,pods: n/a"]')).nativeElement
-        .innerText
-    ).toContain('Memory: 0.0 bytes\nCPU: 0');
-  });
-
-  it('should have usedResource and pendingResource column with detailToggle ON', () => {
-    let service: SchedulerService;
-    service = TestBed.inject(SchedulerService);
-    let appInfo = new AppInfo(
-      'app1',
-      'Memory: 500.0 KB, CPU: 10, pods: 1',
-      'Memory: 0.0 bytes, CPU: 0, pods: n/a',
-      '',
-      1,
-      2,
-      [],
-      2,
-      'RUNNING',
-      []
-    );
-    spyOn(service, 'fetchAppList').and.returnValue(of([appInfo]));
-    component.fetchAppListForPartitionAndQueue('default', 'root');
-    component.detailToggle = true;
-    fixture.detectChanges();
-    const debugEl: DebugElement = fixture.debugElement;
-    expect(
-      debugEl.query(By.css('[data-test="Memory: 500.0 KB,CPU: 10,pods: 1"]')).nativeElement
-        .innerText
+      debugEl.query(By.css('mat-cell.mat-column-usedResource')).nativeElement.innerText
     ).toContain('Memory: 500.0 KB\nCPU: 10\npods: 1');
     expect(
-      debugEl.query(By.css('[data-test="Memory: 0.0 bytes,CPU: 0,pods: n/a"]')).nativeElement
-        .innerText
+      debugEl.query(By.css('mat-cell.mat-column-pendingResource')).nativeElement.innerText
     ).toContain('Memory: 0.0 bytes\nCPU: 0\npods: n/a');
+  });
+
+  it('should copy the allocations URL to clipboard', () => {
+    const debugEl: DebugElement = fixture.debugElement;
+    const copyButton = debugEl.query(By.css('.copy-btn'));
+    const copyButtonSpy = spyOn(component, 'copyLinkToClipboard');
+    copyButton.triggerEventHandler('click', null);
+    expect(copyButtonSpy).toHaveBeenCalled();
   });
 });
