+++ src/app/components/apps-view/apps-view.component.ts
@@ -16,36 +16,25 @@
  * limitations under the License.
  */
 
-import {
-  Component,
-  ComponentRef,
-  ElementRef,
-  OnInit,
-  ViewChild,
-  ViewContainerRef,
-} from '@angular/core';
-import { MatPaginator } from '@angular/material/paginator';
-import { MatSelect, MatSelectChange } from '@angular/material/select';
-import { MatSort } from '@angular/material/sort';
-import { MatTableDataSource } from '@angular/material/table';
+import { Component, OnInit, ViewChild, ElementRef } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
-import { AllocationsDrawerComponent } from '@app/components/allocations-drawer/allocations-drawer.component';
-import { AllocationInfo } from '@app/models/alloc-info.model';
-import { AppInfo } from '@app/models/app-info.model';
-import { ColumnDef } from '@app/models/column-def.model';
-import { DropdownItem } from '@app/models/dropdown-item.model';
-import { PartitionInfo } from '@app/models/partition-info.model';
-import { QueueInfo } from '@app/models/queue-info.model';
-import { EnvconfigService } from '@app/services/envconfig/envconfig.service';
-import { SchedulerService } from '@app/services/scheduler/scheduler.service';
-import { CommonUtil } from '@app/utils/common.util';
+import { MatPaginator } from '@angular/material/paginator';
+import { MatTableDataSource } from '@angular/material/table';
+import { MatSort } from '@angular/material/sort';
+import { MatSelectChange, MatSelect } from '@angular/material/select';
+import { finalize, debounceTime, distinctUntilChanged } from 'rxjs/operators';
 import { NgxSpinnerService } from 'ngx-spinner';
 import { fromEvent } from 'rxjs';
-import { debounceTime, distinctUntilChanged, finalize } from 'rxjs/operators';
-import {
-  loadRemoteModule,
-  LoadRemoteModuleEsmOptions,
-} from '@angular-architects/module-federation';
+
+import { SchedulerService } from '@app/services/scheduler/scheduler.service';
+import { AppInfo } from '@app/models/app-info.model';
+import { AllocationInfo } from '@app/models/alloc-info.model';
+import { ColumnDef } from '@app/models/column-def.model';
+import { CommonUtil } from '@app/utils/common.util';
+import { PartitionInfo } from '@app/models/partition-info.model';
+import { DropdownItem } from '@app/models/dropdown-item.model';
+import { QueueInfo } from '@app/models/queue-info.model';
+import { MatDrawer } from '@angular/material/sidenav';
 
 @Component({
   selector: 'app-applications-view',
@@ -59,12 +48,7 @@ export class AppsViewComponent implements OnInit {
   @ViewChild('allocSort', { static: true }) allocSort!: MatSort;
   @ViewChild('searchInput', { static: true }) searchInput!: ElementRef;
   @ViewChild('queueSelect', { static: false }) queueSelect!: MatSelect;
-
-  @ViewChild('drawerContainer', { read: ViewContainerRef, static: true })
-  drawerContainer!: ViewContainerRef;
-
-  @ViewChild('mfeContainer', { read: ViewContainerRef, static: true })
-  mfeContainer!: ViewContainerRef;
+  @ViewChild('matDrawer', { static: false }) matDrawer!: MatDrawer;
 
   appDataSource = new MatTableDataSource<AppInfo>([]);
   appColumnDef: ColumnDef[] = [];
@@ -83,14 +67,13 @@ export class AppsViewComponent implements OnInit {
   leafQueueSelected = '';
 
   detailToggle: boolean = false;
-  allocationsDrawerComponent: ComponentRef<AllocationsDrawerComponent> | undefined = undefined;
+  allocationsToggle: boolean = false;
 
   constructor(
     private scheduler: SchedulerService,
     private spinner: NgxSpinnerService,
     private activatedRoute: ActivatedRoute,
-    private router: Router,
-    private envConfig: EnvconfigService
+    private router: Router
   ) {}
 
   ngOnInit() {
@@ -101,31 +84,27 @@ export class AppsViewComponent implements OnInit {
     this.appSort.sort({ id: 'submissionTime', start: 'desc', disableClear: false });
 
     this.appColumnDef = [
-      { colId: 'applicationId', colName: 'Application ID', colWidth: 1 },
-      { colId: 'applicationState', colName: 'Application State', colWidth: 1 },
+      { colId: 'applicationId', colName: 'Application ID' },
+      { colId: 'applicationState', colName: 'Application State' },
       {
         colId: 'lastStateChangeTime',
         colName: 'Last State Change Time',
         colFormatter: CommonUtil.timeColumnFormatter,
-        colWidth: 1,
       },
       {
         colId: 'usedResource',
         colName: 'Used Resource',
         colFormatter: CommonUtil.resourceColumnFormatter,
-        colWidth: 2,
       },
       {
         colId: 'pendingResource',
         colName: 'Pending Resource',
         colFormatter: CommonUtil.resourceColumnFormatter,
-        colWidth: 2,
       },
       {
         colId: 'submissionTime',
         colName: 'Submission Time',
         colFormatter: CommonUtil.timeColumnFormatter,
-        colWidth: 1,
       },
     ];
 
@@ -175,31 +154,6 @@ export class AppsViewComponent implements OnInit {
           this.clearQueueSelection();
         }
       });
-
-    this.initializeSidebarComponent(
-      this.envConfig.getAllocationsDrawerComponentRemoteConfig()
-    ).catch(console.error);
-  }
-
-  async initializeSidebarComponent(remoteComponentConfig: LoadRemoteModuleEsmOptions | null) {
-    let component = AllocationsDrawerComponent;
-
-    if (remoteComponentConfig !== null) {
-      const { AllocationsDrawerComponent: allocationsDrawerComponent } =
-        await loadRemoteModule(remoteComponentConfig);
-      component = allocationsDrawerComponent;
-    }
-
-    this.drawerContainer.clear();
-    this.allocationsDrawerComponent = this.drawerContainer.createComponent(component);
-
-    this.allocationsDrawerComponent.setInput('selectedRow', this.selectedRow);
-    this.allocationsDrawerComponent.setInput('allocDataSource', this.allocDataSource);
-    this.allocationsDrawerComponent.setInput('partitionSelected', this.partitionSelected);
-    this.allocationsDrawerComponent.setInput('leafQueueSelected', this.leafQueueSelected);
-    this.allocationsDrawerComponent.instance.removeRowSelection.subscribe(() => {
-      this.removeRowSelection();
-    });
   }
 
   fetchQueuesForPartition(partitionName: string) {
@@ -323,12 +277,10 @@ export class AppsViewComponent implements OnInit {
     } else {
       this.selectedRow = row;
       row.isSelected = true;
+      this.matDrawer.open();
       if (row.allocations) {
         this.allocDataSource.data = row.allocations;
       }
-      this.allocationsDrawerComponent?.setInput('selectedRow', this.selectedRow);
-      this.allocationsDrawerComponent?.setInput('allocDataSource', this.allocDataSource);
-      this.allocationsDrawerComponent?.instance.openDrawer();
     }
   }
 
@@ -346,6 +298,10 @@ export class AppsViewComponent implements OnInit {
 
   isAppDataSourceEmpty() {
     return this.appDataSource.data && this.appDataSource.data.length === 0;
+  }
+
+  isAllocDataSourceEmpty() {
+    return this.allocDataSource.data && this.allocDataSource.data.length === 0;
   }
 
   onClearSearch() {
@@ -421,7 +377,9 @@ export class AppsViewComponent implements OnInit {
     const otherElements = arr.filter(
       (item) => !item.toLowerCase().includes('CPU') && !item.toLowerCase().includes('Memory')
     );
-    return cpuAndMemoryElements.concat(otherElements);
+    const result = cpuAndMemoryElements.concat(otherElements);
+
+    return result;
   }
 
   clearQueueSelection() {
@@ -436,5 +394,22 @@ export class AppsViewComponent implements OnInit {
 
   toggle() {
     this.detailToggle = !this.detailToggle;
+  }
+
+  allocationsDetailToggle() {
+    this.allocationsToggle = !this.allocationsToggle;
+  }
+
+  closeDrawer() {
+    this.matDrawer.close();
+    this.removeRowSelection();
+  }
+
+  copyLinkToClipboard() {
+    const url = window.location.href.split('?')[0];
+    const copyString = `${url}?partition=${this.partitionSelected}&queue=${this.leafQueueSelected}&applicationId=${this?.selectedRow?.applicationId}`;
+    navigator.clipboard
+      .writeText(copyString)
+      .catch((error) => console.error('Writing to the clipboard is not allowed. ', error));
   }
 }
