+++ src/app/models/envconfig.model.ts
@@ -18,6 +18,4 @@
 
 export interface EnvConfig {
   localSchedulerWebAddress: string;
-  moduleFederationRemoteEntry?: string;
-  allocationsDrawerRemoteComponent?: string;
 }
