+++ src/app/testing/mocks.ts
@@ -21,7 +21,6 @@ import { AppInfo } from '@app/models/app-info.model';
 import { CommonUtil } from '@app/utils/common.util';
 
 export const noopFn = () => {};
-export const nullFn = () => null;
 
 export const MockSchedulerService = {
   fetchClusterByName: () => of({}),
@@ -42,7 +41,6 @@ export const MockNgxSpinnerService = {
 
 export const MockEnvconfigService = {
   getSchedulerWebAddress: noopFn,
-  getAllocationsDrawerComponentRemoteConfig: nullFn,
 };
 
 export const MockEventBusService = {
